backend/temporal/pyproject.toml
# Copyright 2023 Canonical Ltd.
# See LICENSE file for licensing details.

[tool.poetry]
name = "msm-worker"
version = "0.1"
description = ""
authors = ["MAAS <maas-devel@lists.launchpad.net>"]
packages = [
    { include = "**/*.py", from = "."  }
]

[tool.poetry.dependencies]
python = "^3.12"
protobuf = "^3.2.0"
PyYAML = "^6.0"
temporal-lib-py = "^1.8.0"
httpx = "^0.28.1"
boto3 = "^1.37.7"
python-json-logger = "^2.0.4"
urllib3 = "^1.26.16"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
black = "^22.8.0"
isort = "^5.10.1"
poethepoet = "^0.16.2"
pytest-asyncio = "^0.19.0"
mypy = "^0.971"

[tool.poe.tasks]
format = [{cmd = "black ."}, {cmd = "isort ."}]
lint = [{cmd = "black --check ."}, {cmd = "isort --check-only ."}, {ref = "lint-types" }]
lint-types = "mypy --check-untyped-defs ."
test = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.mypy]
ignore_missing_imports = true

[tool.basedpyright]
include = [
  "resources",
  "tests",
]
exclude = [
  "**/__pycache__",
  "venv",
  ".ve",
  ".tox",
]
typeCheckingMode = "strict"
reportMissingImports = true
reportUnusedImport = true
reportMissingTypeStubs = false
reportUnusedCallResult = false
pythonVersion = "3.12"
